fix(tutorials): apply deselect on empty on spatial target dispatcher
The Teleport Targets and Free Teleporting tutorial had an issue
where if a teleport point was teleported to and then the free
teleporter was used to teleport away from the selected teleport
point, then the current activated teleport point would not deactivate
and therefore still be hidden.

The code deals with this scenario, but the spatial target dispatcher
needs to have the `Deselect All On Empty Target` option selected to
ensure when the free teleporter is used that the spatial target
dispatcher now has an empty target so should disable any existing
enabled targets.